+++ src/by/it/akhmelev/jd02_02/th2_01_not_syn.java
@@ -1,9 +1,9 @@
 package by.it.akhmelev.jd02_02;
 
-public class Th2_02_volatile_ok {
+public class th2_01_not_syn {
 
     //переменная баланса
-    volatile static int balance=0;
+    static Integer balance=new Integer(0);
 
     //это касса. Просто добавляет в баланс единицу
     static class Cashier extends Thread {
@@ -14,10 +14,8 @@ public class Th2_02_volatile_ok {
         }
         @Override
         public  void run() {
-            //так ПОЧТИ ВСЕГДА будет работать.
-            // volatile - это ГАРАНТИЯ доступности, а не гарантия целостности.
-            int delta = (calc(1));
-            balance+=delta; //но потенциально проблема есть. Тут ТРИ операции, а не одна.
+            //не будет работать. Одновременный доступ.
+                balance += (calc(1));
         }
     }
     //создадим 6666 касс. Каждая добавит по 1. Сколько всего будет?
@@ -27,9 +25,9 @@ public class Th2_02_volatile_ok {
 
         for (int i = 0; i < 6666; i++) {
             new Cashier().start();
-        }
+         }
         //пока потоков больше чем было в начале просто ждем
         while (Thread.activeCount()>thCount) {Thread.sleep(100);}
         System.out.print("Итого:"+balance);
-    }
+     }
 }
