+++ src/by/it/akhmelev/jd02_02/th2_02_volatile_bad.java
@@ -1,9 +1,9 @@
 package by.it.akhmelev.jd02_02;
 
-public class Th2_01_not_syn {
+public class th2_02_volatile_bad {
 
     //переменная баланса
-    static Integer balance=new Integer(0);
+    volatile static Integer balance=0;
 
     //это касса. Просто добавляет в баланс единицу
     static class Cashier extends Thread {
@@ -14,8 +14,8 @@ public class Th2_01_not_syn {
         }
         @Override
         public  void run() {
-            //не будет работать. Одновременный доступ.
-                balance += (calc(1));
+            //не будет работать. volatile - это ГАРАНТИЯ доступности, а не гарантия целостности.
+            balance += (calc(1));
         }
     }
     //создадим 6666 касс. Каждая добавит по 1. Сколько всего будет?
@@ -25,9 +25,9 @@ public class Th2_01_not_syn {
 
         for (int i = 0; i < 6666; i++) {
             new Cashier().start();
-         }
+        }
         //пока потоков больше чем было в начале просто ждем
         while (Thread.activeCount()>thCount) {Thread.sleep(100);}
         System.out.print("Итого:"+balance);
-     }
+    }
 }
