+++ src/by/it/akhmelev/jd02_02/th2_03_with_synchro.java
@@ -1,21 +1,26 @@
 package by.it.akhmelev.jd02_02;
 
-public class Th2_02_volatile_bad {
+public class th2_03_with_synchro {
 
     //переменная баланса
-    volatile static Integer balance=0;
+    static Integer balance=new Integer(0);
 
     //это касса. Просто добавляет в баланс единицу
     static class Cashier extends Thread {
         //создадим видимость расчета
         int calc(int in) {
             int j=0; for (int i = 0; i < 666; i++) {j=j+(int)((Math.sqrt(i)));}
+            balance += in;
             return in;
         }
         @Override
         public  void run() {
-            //не будет работать. volatile - это ГАРАНТИЯ доступности, а не гарантия целостности.
-            balance += (calc(1));
+            synchronized (balance) {
+                //Удивительно, но и это не будет работать
+                balance += (calc(1));
+                //Почему? Потому, что нельзя базировать синхронизатор на изменяемом поле.
+                //Есть шанс потерять монитор, а с ним и синхронизацию
+            }
         }
     }
     //создадим 6666 касс. Каждая добавит по 1. Сколько всего будет?
