+++ src/by/it/akhmelev/jd02_02/th2_03_with_synchro_method.java
@@ -1,9 +1,8 @@
 package by.it.akhmelev.jd02_02;
 
-public class Th2_05_with_synchro_ok {
+public class th2_03_with_synchro_method {
 
     //переменная баланса
-    private final static Integer fakeBalance=0;
     static Integer balance=new Integer(0);
 
     //это касса. Просто добавляет в баланс единицу
@@ -11,14 +10,15 @@ public class Th2_05_with_synchro_ok {
         //создадим видимость расчета
         int calc(int in) {
             int j=0; for (int i = 0; i < 666; i++) {j=j+(int)((Math.sqrt(i)));}
+            balance += in;
             return in;
         }
         @Override
-        public  void run() {
-                synchronized (fakeBalance) {
-                    balance += (calc(1));
-                    //fakeBalance++; //а вот так нельзя, поломаем, поле станет изменяемым
-                }
+        public synchronized void run() {
+            //Удивительно, но и так тоже не будет работать
+            balance += (calc(1));
+            //Почему? Потому, что синхронизатор привязан ко потоку, т.е. к this.
+            //Поэтому он другим не мешает а значит ничего не блокирует
         }
     }
     //создадим 6666 касс. Каждая добавит по 1. Сколько всего будет?
@@ -29,7 +29,7 @@ public class Th2_05_with_synchro_ok {
         for (int i = 0; i < 6666; i++) {
             new Cashier().start();
         }
-        //пока потоков больше, чем было в начале просто ждем
+        //пока потоков больше чем было в начале просто ждем
         while (Thread.activeCount()>thCount) {Thread.sleep(100);}
         System.out.print("Итого:"+balance);
     }
