+++ src/by/it/akhmelev/jd02_02/th2_05_with_synchro_ok.java
@@ -1,8 +1,9 @@
 package by.it.akhmelev.jd02_02;
 
-public class Th2_03_with_synchro {
+public class th2_05_with_synchro_ok {
 
     //переменная баланса
+    private final static Integer fakeBalance=0;
     static Integer balance=new Integer(0);
 
     //это касса. Просто добавляет в баланс единицу
@@ -10,17 +11,14 @@ public class Th2_03_with_synchro {
         //создадим видимость расчета
         int calc(int in) {
             int j=0; for (int i = 0; i < 666; i++) {j=j+(int)((Math.sqrt(i)));}
-            balance += in;
             return in;
         }
         @Override
         public  void run() {
-            synchronized (balance) {
-                //Удивительно, но и это не будет работать
-                balance += (calc(1));
-                //Почему? Потому, что нельзя базировать синхронизатор на изменяемом поле.
-                //Есть шанс потерять монитор, а с ним и синхронизацию
-            }
+                synchronized (fakeBalance) {
+                    balance += (calc(1));
+                    //fakeBalance++; //а вот так нельзя, поломаем, поле станет изменяемым
+                }
         }
     }
     //создадим 6666 касс. Каждая добавит по 1. Сколько всего будет?
@@ -28,10 +26,10 @@ public class Th2_03_with_synchro {
         //Считаем сколько было потоков
         int thCount=Thread.activeCount();
 
-        for (int i = 0; i < 4666; i++) {
+        for (int i = 0; i < 6666; i++) {
             new Cashier().start();
         }
-        //пока потоков больше чем было в начале просто ждем
+        //пока потоков больше, чем было в начале просто ждем
         while (Thread.activeCount()>thCount) {Thread.sleep(100);}
         System.out.print("Итого:"+balance);
     }
