+++ src/by/it/akhmelev/jd02_02/th2_09_deadlock.java
@@ -1,17 +1,17 @@
 package by.it.akhmelev.jd02_02;
 
-public class Th2_09_deadlock {
-    private static class Friend {
+public class th2_09_deadlock {
+    static class Friend {
         private final String name;
-        Friend(String name) {this.name = name;}
+        public Friend(String name) {this.name = name;}
         public String getName() {return this.name;}
 
-        synchronized void bow(Friend bower) {
+        public synchronized void bow(Friend bower) {
             System.out.format("%s: %s" + " поклонился мне!%n", this.name, bower.getName());
             bower.bowBack(this);
         }
 
-        synchronized void bowBack(Friend bower) {
+        public synchronized void bowBack(Friend bower) {
             System.out.format("%s: %s" + " стал прямо!%n",    this.name, bower.getName());
         }
     }
